move report endpoint
