+++ geotrek/feedback/views.py
@@ -1,32 +1,17 @@
-import os
-
 from crispy_forms.helper import FormHelper
 from django.conf import settings
-from django.contrib.auth import get_user_model
-from django.contrib.contenttypes.models import ContentType
 from django.contrib.gis.db.models.functions import Transform
-from django.core.exceptions import ValidationError
-from django.core.files import File
-from django.core.mail import send_mail
 from django.db.models import CharField, F, Value
 from django.db.models.functions import Concat
 from django.urls.base import reverse
 from django.utils.translation import get_language
 from django.utils.translation import gettext as _
-from django.views.generic.list import ListView
 from mapentity import views as mapentity_views
-from PIL import Image
 from rest_framework.authentication import (BasicAuthentication,
                                            SessionAuthentication)
-from rest_framework.decorators import action
-from rest_framework.parsers import FormParser, MultiPartParser
-from rest_framework.permissions import AllowAny, IsAuthenticated
-from rest_framework.response import Response
-from rest_framework.views import APIView
+from rest_framework.permissions import IsAuthenticated
 
-from geotrek.common.mixins.api import APIViewSet
 from geotrek.common.mixins.views import CustomColumnsMixin
-from geotrek.common.models import Attachment, FileType
 from geotrek.common.viewsets import GeotrekMapentityViewSet
 
 from . import models as feedback_models
@@ -92,41 +77,6 @@ class ReportFormatList(mapentity_views.MapEntityFormat, ReportList):
         return super().get_context_data(**kwargs)
 
 
-class CategoryList(mapentity_views.JSONResponseMixin, ListView):
-    model = feedback_models.ReportCategory
-
-    def get_context_data(self, **kwargs):
-        return [{"id": c.id, "label": c.label} for c in self.object_list]
-
-
-class FeedbackOptionsView(APIView):
-    permission_classes = [
-        AllowAny,
-    ]
-
-    def get(self, request, *args, **kwargs):
-        categories = feedback_models.ReportCategory.objects.all()
-        cat_serializer = feedback_serializers.ReportCategorySerializer(
-            categories, many=True
-        )
-        activities = feedback_models.ReportActivity.objects.all()
-        activities_serializer = feedback_serializers.ReportActivitySerializer(
-            activities, many=True
-        )
-        magnitude_problems = feedback_models.ReportProblemMagnitude.objects.all()
-        mag_serializer = feedback_serializers.ReportProblemMagnitudeSerializer(
-            magnitude_problems, many=True
-        )
-
-        options = {
-            "categories": cat_serializer.data,
-            "activities": activities_serializer.data,
-            "magnitudeProblems": mag_serializer.data,
-        }
-
-        return Response(options)
-
-
 class ReportCreate(mapentity_views.MapEntityCreate):
     model = feedback_models.Report
     form_class = ReportForm
@@ -182,76 +132,3 @@ class ReportViewSet(GeotrekMapentityViewSet):
                 self.request.user.pk if settings.SURICATE_WORKFLOW_ENABLED else ''
             )
         return geojson_lookup
-
-
-class ReportAPIViewSet(APIViewSet):
-    queryset = feedback_models.Report.objects.existing()\
-                              .select_related("activity", "category", "problem_magnitude", "status", "related_trek")\
-                              .prefetch_related("attachments")
-    parser_classes = [FormParser, MultiPartParser]
-    serializer_class = feedback_serializers.ReportAPISerializer
-    geojson_serializer_class = feedback_serializers.ReportAPIGeojsonSerializer
-    authentication_classes = []
-    permission_classes = [AllowAny]
-
-    def get_queryset(self):
-        queryset = super().get_queryset()
-        return queryset.select_related(
-            "activity", "category", "problem_magnitude", "status", "related_trek"
-        )
-
-    @action(detail=False, methods=["post"])
-    def report(self, request, lang=None):
-        response = super().create(request)
-        creator, created = get_user_model().objects.get_or_create(
-            username="feedback", defaults={"is_active": False}
-        )
-        for file in request._request.FILES.values():
-            attachment = Attachment(
-                filetype=FileType.objects.get_or_create(type=settings.REPORT_FILETYPE)[
-                    0
-                ],
-                content_type=ContentType.objects.get_for_model(feedback_models.Report),
-                object_id=response.data.get("id"),
-                creator=creator,
-                attachment_file=file,
-            )
-            name, extension = os.path.splitext(file.name)
-            try:
-                attachment.full_clean()  # Check that file extension and mimetypes are allowed
-            except ValidationError as e:
-                logger.error(f"Invalid attachment {name}{extension} for report {response.data.get('id')} : " + str(e))
-            else:
-                try:
-                    # Reencode file to bitmap then back to jpeg lfor safety
-                    if not os.path.exists(f"{settings.TMP_DIR}/report_file/"):
-                        os.mkdir(f"{settings.TMP_DIR}/report_file/")
-                    tmp_bmp_path = os.path.join(f"{settings.TMP_DIR}/report_file/", f"{name}.bmp")
-                    tmp_jpeg_path = os.path.join(f"{settings.TMP_DIR}/report_file/", f"{name}.jpeg")
-                    Image.open(file).save(tmp_bmp_path)
-                    Image.open(tmp_bmp_path).save(tmp_jpeg_path)
-                    with open(tmp_jpeg_path, 'rb') as converted_file:
-                        attachment.attachment_file = File(converted_file, name=f"{name}.jpeg")
-                        attachment.save()
-                    os.remove(tmp_bmp_path)
-                    os.remove(tmp_jpeg_path)
-                except Exception as e:
-                    logger.error(f"Failed to convert attachment {name}{extension} for report {response.data.get('id')}: " + str(e))
-
-        if settings.SEND_REPORT_ACK and response.status_code == 201:
-            send_mail(
-                _("Geotrek : Signal a mistake"),
-                _(
-                    """Hello,
-
-We acknowledge receipt of your feedback, thank you for your interest in Geotrek.
-
-Best regards,
-
-The Geotrek Team
-http://www.geotrek.fr"""
-                ),
-                settings.DEFAULT_FROM_EMAIL,
-                [request.data.get("email")],
-            )
-        return response
